+++ public/css/mobile.css
@@ -1,0 +1,243 @@
+div {
+}
+
+body {
+    margin: 0;
+    padding: 0;
+    margin-left: 0;
+    margin-right: 0;
+    background-color: #4c64ba;
+    word-break: keep-all;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.HeaderTitle {
+    background-color: #fff;
+    background-size: auto;
+    top: 0px;
+    text-align: center;
+    line-height: 60px; 
+    height: 60px;
+    font-size: 25px;
+    font-family: 'Montserrat', sans-serif;
+    font-weight: 900;
+    font-style: italic;
+}
+
+.MainBox {
+    text-align: center;
+    display: block;
+    background: #fff;
+    width: 90%;
+    margin: 20% auto;
+    box-sizing: border-box;
+    padding: 10% 8%;
+    overflow: hidden;
+    border-radius: 20px;
+    left: 5%;
+}
+
+.MainSubBox {
+    background-color: #DBDEE2;
+    padding: 10px;
+    margin-top: 15px;
+    width: 100%;
+    font-size: 14px;
+    font-weight: 600;
+    overflow: hidden;
+    border-radius: 10px;
+    box-sizing: border-box;
+}
+
+.LoginBox {
+    margin-top: 25px;
+    padding: 5px 15px 25px 15px;
+    border: 1px solid #eee;
+    border-radius: 10px;
+}
+
+.BoxTextTitle {
+    position: relative;
+    font-size: 21px;
+    font-weight: 600;
+    letter-spacing:-0.05em;
+    width: 100%;
+    display: block;
+    padding-bottom: 15px;
+    margin-bottom: 15px;
+    border-bottom: 1px solid #ededed;
+}
+
+.BoxTextConnect {
+    position: relative;
+    font-size: 14px;
+    font-weight: 500;
+    color:#888;
+    display: inline-block;
+    line-height: 165%;
+}
+
+.SubBoxText {
+    position: relative;
+    line-height: 150%;
+}
+
+.BoxTextCode {
+    position: relative;
+    font-size: 16px;
+    font-weight: 600;
+    letter-spacing:-0.05em;
+    width: 100%;
+    display: block;
+    text-align: left;
+    margin: 15px 0 5px 0;
+}
+
+
+.InputBox {
+    width: 100%;
+}
+
+
+.MainButton {
+    font-size: 16px;
+    color: #fff;
+    padding: 10px;
+    margin-top: 20px;
+    width: 100%;
+    background-color: #ff9d0a;
+    border-radius: 5px;
+    border: 0px;
+    font-weight: 600;
+}
+
+.GuideTextTitle {
+    position: relative;
+    font-size: 20px;
+    font-weight: 700;
+    letter-spacing:-0.05em;
+    width: 100%;
+    display: block;
+    padding-bottom: 15px;
+    margin-bottom: 15px;
+    border-bottom: 1px solid #ededed;
+}
+
+.GuideStartText {
+    position: relative;
+    display: block;
+    line-height: 150%;
+}
+
+.GuideCameraText {
+    position:relative;
+    display: block;
+    font-weight: 500;
+    line-height: 150%;
+}
+
+.GuideAllowTitle {
+    color: #777;
+    font-size: 18px;
+    font-weight: 400;
+}
+
+.GuideAllowText {
+    color: #999;
+    font-size: 13px;
+    display: block;
+    margin-top: -5px;
+    margin-bottom: 20px;
+}
+
+.GuideAllowBox {
+    margin-top: 20px;
+}
+
+.GuideBox {
+    padding-bottom: 15px;
+    margin: 15px 0;
+}
+
+.GuideImageBox {
+    width: 100%;
+    background-color: #DBDEE2;
+    padding: 20px 0;
+    margin: 10px 0;
+    border-radius: 5px;
+}
+
+.GuideButton {
+    font-size: 16px;
+    color: #fff;
+    padding: 10px;
+    margin-top: 20px;
+    width: 100%;
+    background-color: #ff9d0a;
+    border-radius: 5px;
+    border: 0px;
+    font-weight: 600;
+}
+
+.CameraTitle {
+    color: #888;
+    font-size: 16px;
+    display: block;
+    margin-top: -5px;
+    margin-bottom: 20px;
+}
+
+.CameraBox {
+    padding-bottom: 15px;
+}
+
+.TestButton {
+    font-size: 16px;
+    color: #fff;
+    padding: 10px;
+    margin-top: 20px;
+    width: 100%;
+    background-color: #ff9d0a;
+    border-radius: 5px;
+    border: 0px;
+    font-weight: 600;
+}
+
+.TestTitle {
+    color: #28a515;
+    font-size: 16px;
+    font-weight: 600;
+    display: block;
+    padding-bottom: 5px;
+}
+
+.TestText {
+    color: #28a515;
+    font-size: 13px;
+}
+
+.TestBox {
+    width: 100%;
+    background-color: #ddffd8;
+    border-radius: 5px;
+    padding: 15px 0;
+    margin-top: 15px;
+}
+
+.RTC {
+    margin-top: 15px;
+    overflow: hidden;
+    border-radius: 10px;
+    border: 1px solid #aaa;
+    width: 100%;
+    height: 100%;
+}
+
+.Footer {
+    background-color: white;
+    position: fixed;
+    background-size: auto;
+    width: 100%;
+    padding-top: 70%;
+    height: 60px;
+}
