+++ src/Components/Mobile/Guide.js
@@ -1,12 +1,9 @@
 import React, { useState, useEffect } from "react";
-import { HashRouter, Link, useHistory } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
 import { Cookies, useCookies } from "react-cookie";
 
-import Header from './Header';
-import Sample1 from '../media/sample1.png';
-import Sample2 from '../media/sample2.png';
-import Sample3 from '../media/sample3.png';
-import "../css.css";
+import Header from './MobileHeader';
+
 
 
 const Guide = () => {
@@ -62,7 +59,7 @@ const Guide = () => {
                         [카메라, 마이크 권한 허용 안내]
                     </span>
                     <div className="GuideImageBox">
-                        <img src={Sample1} />
+                        <img src="./img/guide1.png" />
                     </div>
                 </div>
                 <div className="GuideAllowBox">
@@ -70,24 +67,22 @@ const Guide = () => {
                         [올바른 촬영 방법]
                     </span>
                     <div className="GuideImageBox">
-                        <img src={Sample2} />
+                        <img src="./img/guide2.png" />
                     </div>
                     <span className="GuideAllowText">
                         거치대나 물건 등을 사용하여 최소 1m 정도 떨어진 곳에 휴대폰을 고정시킵니다. 
                     </span>
                     <div className="GuideImageBox">
-                        <img src={Sample3} />
+                        <img src="./img/guide3.png" />
                     </div>
                     <span className="GuideAllowText">
                         얼굴과 손, 모니터 화면이 동시에 나오도록 카메라 화면을 조정하세요.
                     </span>
                 </div>
                 <div>
-                    <HashRouter>
                     <Link to="/Camera">
                     <button className="GuideButton" >촬영 시작</button>
                     </Link>
-                    </HashRouter>
                 </div>
             </div>
         </div>
