+++ src/Components/Mobile/MobileHeader.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import "../css.css";
+
 
 function Header() {
     return (
