--- src/components/Footer.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "../css.css";
-
-function Footer() {
-    return (
-        <div>
-            <footer classname="Footer">
-                dfdf
-            </footer>
-        </div>
-    );
-}
-
-export default Footer;
