+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Router from './components/Router';
+import App from './App';
 
 
 ReactDOM.render(
-  <Router />,
+  <App />,
   document.getElementById('root')
 );
